refactor: strutil 라이브러리 정리
