+++ src/util/strutil/conversion.cpp
@@ -1,0 +1,13 @@
+#include <fstream>
+#include <string>
+#include <util/strutil/strutil.hpp>
+
+namespace util {
+using std::string;
+
+template <>
+string to_string(const string& t) {
+  return t;
+}
+
+}  // namespace util
