+++ src/util/strutil/join.tpp
@@ -1,0 +1,9 @@
+#ifndef UTIL_STRUTIL_JOIN_TPP
+#define UTIL_STRUTIL_JOIN_TPP
+
+#include <util/strutil/join.hpp>
+
+namespace util {
+}  // namespace util
+
+#endif // UTIL_STRUTIL_JOIN_TPP
