+++ src/util/strutil/read.cpp
@@ -1,6 +1,7 @@
 #include <fstream>
+#include <iostream>
+#include <sstream>
 #include <string>
-#include <util/strutil/strutil.hpp>
 
 namespace util {
 using std::string;
@@ -11,11 +12,6 @@ string read_text(const string& filename) {
 
   buffer << ifs.rdbuf();
   return buffer.str();
-}
-
-template <>
-string to_string(const string& t) {
-  return t;
 }
 
 }  // namespace util
